requirements.txt
absl-py==1.2.0
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
anyio==3.6.1
appnope==0.1.3
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
backports.cached-property==1.0.1
beautifulsoup4==4.11.1
bleach==5.0.0
boto3==1.23.8
botocore==1.26.8
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coloredlogs==15.0.1
cycler==0.11.0
dacite==1.6.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
entrypoints==0.4
eventlet==0.33.1
executing==0.8.3
fastjsonschema==2.15.3
fiftyone==0.16.1
fiftyone-brain==0.8.2
fiftyone-db==0.3.0
Flask==2.1.2
flatbuffers==2.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
glob2==0.7
google-auth==2.9.1
google-auth-oauthlib==0.4.6
graphql-core==3.1.7
greenlet==1.1.2
grpcio==1.47.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
humanfriendly==10.0
hydra-core==1.2.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
imageio==2.19.2
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
json5==0.9.8
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
kaleido==0.2.1
kiwisolver==1.4.2
lightly==1.2.24
lightly-utils==0.0.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mongoengine==0.20.0
motor==2.5.1
mpmath==1.2.1
multidict==6.0.2
nbclassic==0.3.7
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
ndjson==0.3.1
nest-asyncio==1.5.5
networkx==2.8.2
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.4
oauthlib==3.2.0
omegaconf==2.2.2
onnx==1.12.0
onnxruntime==1.12.0
opencv-contrib-python==4.5.5.64
opencv-python==4.6.0.66
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
patool==1.12
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
plotly==4.14.3
pprintpp==0.4.0
priority==2.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.12.0
pymongo==3.12.3
pyparsing==3.0.9
pyrsistent==0.18.1
pyspark==3.2.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.6.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.0.0
redis==4.3.4
requests==2.27.1
requests-oauthlib==1.3.1
retrying==1.3.3
rfc3986==1.5.0
rsa==4.9
s3transfer==0.5.2
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
sentinel==0.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sse-starlette==0.10.3
sseclient-py==1.7.2
stack-data==0.2.0
starlette==0.16.0
strawberry-graphql==0.96.0
sympy==1.10.1
tabulate==0.8.9
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
thop==0.1.1.post2207130030
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
toml==0.10.2
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.9.3
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
universal-analytics-python3==1.1.1
urllib3==1.26.9
voxel51-eta==0.7.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
xmltodict==0.13.0
yarl==1.7.2
